+++ ptai-ee-tools-java/generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/ptaiserver/utils/LoggingInterceptor.java
@@ -1,9 +1,8 @@
 package com.ptsecurity.appsec.ai.ee.utils.ci.integration.ptaiserver.utils;
 
+import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import okhttp3.Interceptor;
-import okhttp3.Request;
-import okhttp3.Response;
+import okhttp3.*;
 import okio.Buffer;
 import okio.BufferedSource;
 
@@ -16,25 +15,46 @@ public class LoggingInterceptor implements Interceptor {
         Request request = chain.request();
 
         long requestTime = System.nanoTime();
-        log.trace(String.format("Sending request to %s", request.url()));
-        log.trace(String.format("Request headers: %s", request.headers()));
+        log.trace("Sending request to {}", request.url());
+        log.trace("Request headers: {}", request.headers());
         if (null != request.body())
-            log.trace(String.format("Request body: %s", request.body().toString()));
+            traceBody(request.headers(), request.body());
 
         Response response = chain.proceed(request);
         long responseTime = System.nanoTime();
         log.trace(String.format("Received %d response for %s in %.1fms",
                 response.code(), response.request().url(), (responseTime - requestTime) / 1e6d));
 
-        log.trace(String.format("Response headers: %s", response.headers()));
+        log.trace("Response headers: {}", response.headers());
         if (null != response.body()) {
             BufferedSource source = response.body().source();
             source.request(Long.MAX_VALUE); // Buffer the entire body.
             Buffer buffer = source.getBuffer();
-            // TODO: Add buffer size verification
             String bufferData = buffer.clone().readString(StandardCharsets.UTF_8).toString();
-            log.trace(String.format("Response body: %s", bufferData));
+            if (5 * 1024 >= bufferData.length()) {
+                log.trace("Response body: {}", bufferData);
+            } else
+                log.trace("Response body skipped as it {} bytes long", bufferData.length());
         }
         return response;
+    }
+
+    private final boolean encodingUnknown(@NonNull final Headers headers) {
+        final String encoding = headers.get("Content-Encoding");
+        return !("identity".equalsIgnoreCase(encoding) || "gzip".equalsIgnoreCase(encoding));
+    }
+
+    protected void traceBody(
+            @NonNull final Headers headers,
+            @NonNull final RequestBody body) throws IOException {
+        long contentLength = body.contentLength();
+        String bodySize = -1L != body.contentLength() ? contentLength + " byte" : "unknown";
+        log.trace("Request body size: {}", bodySize);
+
+        if (!"application/json".equalsIgnoreCase(headers.get("Content-Type"))) return;
+
+        Buffer buffer = new Buffer();
+        body.writeTo(buffer);
+        log.trace("Request body: {}", buffer.readString(StandardCharsets.UTF_8));
     }
 }